Alterei o tipo da matricula em AlunoConverter, AlunoService, AlunoDAO de Long para Integer, para poder recuperar os dados do aluno na hora de editar.

+++ ProjetoBiblioteca/src/br/edu/ifpb/mt/daca/dao/AlunoDAO.java
@@ -26,7 +26,7 @@ public class AlunoDAO extends DAO {
 		}
 	}
 
-	public Aluno buscar(Long matriculaAluno) throws BibliotecaException {
+	public Aluno buscar(Integer matriculaAluno) throws BibliotecaException {
 		EntityManager em = getEntityManager();
 		Aluno resultado = null;
 		try {
